requirements.txt
dash
odoorpc
gunicorn
python-dotenv
dash-mantine-components